[#121]refactoring: Admin 컨트롤러 커맨드 쿼리로 분리
